+++ source/apps/tmdb/models.py
@@ -45,6 +45,9 @@ class Progress(BaseModel):
     show_status = models.CharField(
         max_length=16, choices=SHOW_STATUS_CHOICES, verbose_name="show status"
     )
+    show_genres = ArrayField(
+        models.CharField(max_length=32), default=list, verbose_name="show genres"
+    )
     show_languages = ArrayField(
         models.CharField(max_length=8), default=list, verbose_name="show languages"
     )
@@ -156,6 +159,7 @@ class Progress(BaseModel):
         self.show_name = self.show.name
         self.show_poster_path = self.show.poster_path
         self.show_status = self.show.status_value
+        self.show_genres = self.show.genres
         self.show_languages = self.show.languages
         self.last_aired_season, self.last_aired_episode = self.show.last_aired_episode